requirements.txt
streamlit~=1.10.0
Pillow~=9.1.1
django-qrcode~=0.3
opencv-python==4.6.0.66
qrcode~=7.3.1
pandas~=1.4.2
numpy~=1.22.4